Replace Slack reference with Discord
PWL migrated off of Slack some time ago and the Slack invite app on
Heroku no longer works. This change updates the README to include an
invite to the PWL Discord server. Since Discord allows invites to land
in specific channels, I've generated a non-expiring invitation to the
`#startingapwl` channel for this context.